Switch `HealthCheck` to TCs
