+++ apps/monolith/test/integration/ContosoUniversity.Mvc.IntegrationTests/HealthCheck/InfraExistsTests.cs
@@ -1,0 +1,41 @@
+namespace ContosoUniversity.Mvc.IntegrationTests.HealthCheck;
+
+using System.Net.Http;
+using System.Net.Http.Json;
+using System.Threading.Tasks;
+
+using FluentAssertions;
+
+using HealthChecks.UI.Core;
+
+using Xunit;
+
+public class InfraExistsTests :
+    IClassFixture<TestsConfiguration>,
+    IClassFixture<CustomWebApplicationFactory>
+{
+    private readonly HttpClient _httpClient;
+
+    public InfraExistsTests(
+        TestsConfiguration config,
+        CustomWebApplicationFactory factory)
+    {
+        factory.ClientOptions.BaseAddress = config.BaseAddressHttpsUrl;
+        _httpClient = factory.CreateClient();
+    }
+
+    [Theory]
+    [InlineData("/health/readiness")]
+    [InlineData("/health/liveness")]
+    public async Task Health_ReturnsHealthy(string healthUrl)
+    {
+        HttpResponseMessage response = await _httpClient.GetAsync(healthUrl);
+
+        response.Should().BeSuccessful();
+        response.Content.Headers.ContentType?.ToString().Should().Be("application/json");
+
+        var report = await response.Content.ReadFromJsonAsync<UIHealthReport>(JsonSerializerOptions.HealthChecks);
+
+        report.ShouldBeHealthy();
+    }
+}
